Make upsert permission more resilient by dropping invalid inheritance references instead of throwing the error upwards
